UBUNTU: [Config] HOTPLUG_PCI_ACPI=y
BugLink: http://bugs.launchpad.net/bugs/1479031

SRU Justification:

Impact: pci hotplug events are lost, and pci devices are ignored
Fix: build pci hotplug driver into kernel
Testcase: start virtual guest and immediately hotplug a pci drive

In -precise, the pci hotplug driver is compiled as a module, but the pci
subsystem is built-in.  This significant amount of time (several seconds)
between when the pci devices are enumerated, and when the pci hotplug
driver begins listening for pci hotplug events (several seconds).  Any
pci devices hotplugged in that interval are ignored.  For virtual guests,
it is possible to hit this window if a drive is added to the virtual guest
immediately after starting it.

This simply compiles the pci hotplug driver into the kernel, which reduces
the interval where pci hotplug events may be missed down to a fraction of
a second (around ~1/4 second).  This isn't 100% safe, but it's far better
than it was with pci hotplug compiled as a module.

Note that commit 3b63aa essentially fixes this, by directly initializing
pci hotplug from the pci subsystem initialization; so this problem exists
only in -precise.

Signed-off-by: Dan Streetman <dan.streetman@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>